otp시스템

+++ Project/src/server/server.java
@@ -4,7 +4,7 @@ import java.net.*;
 import java.util.*;
 
 public class server {
-	private static Map<String,Integer> ipList = new HashMap<>();
+	private static Map<String,Long> ipList = new HashMap<>();
 	public static void main(String[] args) {
 		try(ServerSocket server = new ServerSocket(20000)){
 			System.out.println("서버 기동");
